Using context notifier instead of timer.
All registered questions and answers with a context (i.e. published
RRs and running queries) will be attached to a contextNotifier. If
a context is closed we will get a notification.

Currently this will trigger a complete check of times (and contexts)
on all questions and answers and expire those that have been closed
which typically should be only the closed context.