Using timer for checking records.
When the TTL is checked for records the "next time"
an event will occurr for a record will be retrieved.
These will be sorted to get the earliest time that hasn't
already passed. This time will be used to set a timer
to wake dnssd goroutine for the next TTL scan.